pyproject.toml
[project]
name = "agentcortex"
version = "0.1.0"
description = "AgentCortex: A comprehensive RAG (Retrieval-Augmented Generation) pipeline with PDF text extraction, vector storage, and LangChain integration"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core dependencies
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-huggingface>=0.0.1",
    
    # PDF processing
    "pypdf2>=3.0.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.9.0",
    
    # Text processing and chunking
    "nltk>=3.8.0",
    "tiktoken>=0.5.0",
    
    # Vector storage and embeddings
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    
    # Hugging Face tools
    "transformers>=4.30.0",
    "evaluate>=0.4.0",
    "datasets>=2.14.0",
    
    # Toxicity filtering
    "detoxify>=0.5.0",
    "torch>=2.0.0",
    
    # Language detection
    "langdetect>=1.0.9",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
gpu = [
    "bitsandbytes>=0.41.0",
]
nlp = [
    "spacy>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
